[ai-form-recognizer] Rename TrainingPollOperationState.
